Don't rely on config filter when sourcing train.R
In thinking through the workflow some more I realized that we can't rely on our config filter to ever be active when the user sources train.R (how could it be unless there is some R code that enables it).

This had a couple of implications:

- The download of local copies of gs_data needs to be explicit rather than implicit in the config filter.

- The job_dir needs to be provided explicitly in the config (note that sourcing train.R was actually broken until I made that change).

As regards permissions on the call to gs_data, I actually had no trouble with that in my setup. I think for demos like this we should store test data in buckets that don't require auth (that way everyone can run train.R locally with no gcloud setup).